# api/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the FAIMS Conductor server API documentation.",
    "version": "1.0.0",
    "title": "FAIMS3 API",
    "termsOfService": "http://faims.edu.au/terms/",
    "contact": {"email": "help@faims.edu.au"},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [{"url": "/api"}],
  "components": {
    "securitySchemes": {
      "Auth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "responses": {
      "UnauthorizedError": {"description": "Access token is missing or invalid"}
    }
  },
  "paths": {
...
    "/teams/": {
      "get": {
        "summary": "Get a list of teams",
        "description": "Gets a list of all teams from the teams DB",
        "tags": ["Teams"],
        "security": [{"Auth": []}],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/GetListTeamsResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      },
      "post": {
        "summary": "Create a new team",
        "description": "Creates a new team. Requires authentication.",
        "tags": ["Teams"],
        "security": [{"Auth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/definitions/PostCreateTeamInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/PostCreateTeamResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      }
    },
    "/teams/{id}": {
      "get": {
        "summary": "Get a team by ID",
        "tags": ["Teams"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "security": [{"Auth": []}],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/GetTeamByIdResponse"}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing team",
        "tags": ["Teams"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "security": [{"Auth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/definitions/PutUpdateTeamInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/PutUpdateTeamResponse"}
              }
            }
          }
        }
      }
    },
...
    "/invites/team/{teamId}": {
      "get": {
        "summary": "Get all team invites",
        "tags": ["Invites"],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "security": [{"Auth": []}],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/GetTeamInvitesResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      },
      "post": {
        "summary": "Create a team invite",
        "description": "Creates a new invite for a team. Requires appropriate permissions.",
        "tags": ["Invites"],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"},
            "description": "Team identifier"
          }
        ],
        "security": [{"Auth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/definitions/PostCreateInviteInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/PostCreateTeamInviteResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      }
    },
...
  },
  "definitions": {
...
    "GetListTeamsResponse": {
      "type": "object",
      "properties": {
        "teams": {
          "type": "array",
          "items": {"$ref": "#/definitions/TeamDocument"}
        }
      }
    },
    "GetTeamByIdResponse": {"$ref": "#/definitions/TeamDocument"},
    "PostCreateTeamInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "minLength": 1,
          "description": "Name of the team"
        },
        "description": {
          "type": "string",
          "description": "Description of the team"
        }
      },
      "required": ["name"]
    },
    "PostCreateTeamResponse": {"$ref": "#/definitions/TeamDocument"},
    "PutUpdateTeamInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "minLength": 1,
          "description": "Name of the team"
        },
        "description": {
          "type": "string",
          "description": "Description of the team"
        }
      }
    },
    "PutUpdateTeamResponse": {"$ref": "#/definitions/TeamDocument"},
    "TeamDocument": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the team"
        },
        "_rev": {
          "type": "string",
          "description": "CouchDB revision identifier"
        },
        "name": {"type": "string", "description": "Name of the team"},
        "description": {
          "type": "string",
          "description": "Description of the team"
        },
        "createdBy": {
          "type": "string",
          "description": "User ID who created the team"
        },
        "createdAt": {
          "type": "number",
          "description": "Unix timestamp when team was created"
        },
        "updatedAt": {
          "type": "number",
          "description": "Unix timestamp when team was last updated"
        }
      },
      "required": ["_id", "name", "createdBy", "createdAt", "updatedAt"]
    },
...
    "GetTeamInvitesResponse": {
      "type": "object",
      "properties": {
        "invites": {
          "type": "array",
          "items": {"$ref": "#/definitions/TeamInviteDocument"}
        }
      }
    },
...
    "TeamInviteDocument": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the invite"
        },
        "_rev": {
          "type": "string",
          "description": "CouchDB revision identifier"
        },
        "teamId": {"type": "string", "description": "Team identifier"},
        "email": {
          "type": "string",
          "format": "email",
          "description": "Email address of the invitee"
        },
        "role": {
          "type": "string",
          "description": "Role to be assigned to the user in the team"
        },
        "invitedBy": {
          "type": "string",
          "description": "User ID who created the invite"
        },
        "createdAt": {
          "type": "number",
          "description": "Unix timestamp when invite was created"
        },
        "expiresAt": {
          "type": "number",
          "description": "Unix timestamp when invite expires"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "expired", "revoked"],
          "description": "Status of the invite"
        }
      },
      "required": [
        "_id",
        "teamId",
        "email",
        "role",
        "invitedBy",
        "createdAt",
        "status"
      ]
    },
    "PostCreateInviteInput": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string",
          "description": "Role to assign to the invited user"
        },
        "name": {
          "type": "string",
          "description": "Name/email of the person being invited"
        },
        "expiry": {
          "type": "number",
          "description": "Unix timestamp when invite expires (optional)"
        },
        "uses": {
          "type": "number",
          "description": "Number of times the invite can be used (optional)"
        }
      },
      "required": ["role", "name"]
    },
...
    "PostCreateTeamInviteResponse": {
      "$ref": "#/definitions/TeamInviteDocument"
    },
...
  }
}